+++ doc/sphinx/src/news.rst
@@ -61,6 +61,7 @@ New features
 
 * Logging
 
+  * ``flog::config`` is the collection of FLOG options that can be changed at runtime.
   * FLOG configuration can be set at runtime via new command-line options ``--flog-color``, ``--flog-serialization-interval``,
     and ``--flog-strip-level``.